+++ instrumentation/github.com/gin-gonic/gin/otelgin/test/gintrace_test.go
@@ -251,3 +251,57 @@ func TestWithFilter(t *testing.T) {
 		assert.Len(t, sr.Ended(), 1)
 	})
 }
+
+func TestRecordPanic(t *testing.T) {
+	recoveryMiddleware := func(c *gin.Context) {
+		// Ensure panics are recovered and don't crash tests/are logged to stdout.
+		defer func() {
+			_ = recover()
+		}()
+		c.Next()
+	}
+
+	testCases := []struct {
+		name             string
+		expectStackTrace bool
+	}{
+		{
+			name:             "should record stack trace",
+			expectStackTrace: true,
+		},
+		{
+			name:             "should not record stack trace",
+			expectStackTrace: false,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			sr := tracetest.NewSpanRecorder()
+			provider := sdktrace.NewTracerProvider()
+			provider.RegisterSpanProcessor(sr)
+			router := gin.New()
+			router.Use(recoveryMiddleware, otelgin.Middleware("potato", otelgin.WithTracerProvider(provider), otelgin.WithRecordPanicStackTrace(tc.expectStackTrace)))
+			router.GET("/user/:id", func(c *gin.Context) { panic("corn") })
+			router.ServeHTTP(httptest.NewRecorder(), httptest.NewRequest("GET", "/user/123", nil))
+
+			require.Len(t, sr.Ended(), 1, "should emit a span")
+			span := sr.Ended()[0]
+			assert.Equal(t, span.Status().Code, codes.Error, "should set Error status for panics")
+			require.Len(t, span.Events(), 1, "should emit an event")
+			event := span.Events()[0]
+			assert.Equal(t, event.Name, "exception")
+
+			var foundStackTrace bool
+
+			for _, attr := range event.Attributes {
+				if attr.Key == "exception.stacktrace" {
+					foundStackTrace = true
+					break
+				}
+			}
+
+			assert.Equal(t, tc.expectStackTrace, foundStackTrace, "should record a stack trace")
+		})
+	}
+}